Factorize initialization in tools, make display optional.
